Setup calender event component
